Remove old search
