+++ website/src/components/index.js
@@ -7,4 +7,3 @@ export { Sidebar } from './Sidebar';
 export { SidebarNav } from './SidebarNav';
 export { Search } from './Search';
 export { SiteMeta } from './SiteMeta';
-export { AlgoliaSearch } from './AlgoliaSearch';
